#1700: extract remaining ShowText branches to topic files
Move the last still-inline ShowText branches out of server_show.go:

- class-of-service, vlans, ipv6-router-advertisement ->
  server_show_interfaces_text.go
- test-policy:, firewall-filter: -> server_show_firewall.go
- test-zone: -> server_show_zones_text.go
- forwarding-options, forwarding-options-port-mirroring ->
  server_show_forwarding.go
- event-options -> server_show_events.go
- login, internet-options, root-authentication, core-dumps, task,
  buffers, buffers-detail, backup-router -> server_show_system.go

Bodies moved verbatim (&buf -> passed-in *strings.Builder); prefix
handlers (class-of-service/test-policy/test-zone/firewall-filter)
return the response directly. No new files: every helper lands in an
existing domain file; largest receiver is security_text at 834 LOC,
all well under the 1500 floor.

server_show.go drops to 535 LOC. ShowText golden guard passes
byte-identical; full pkg/grpcapi suite green.
